Move ClientPool back to httplib2
